Moved code to the src folder
