+++ Software/Arduino code/OpenAstroTracker/OpenAstroTracker.ino
@@ -1,18 +1,3 @@
 // Configuration is now done in Configuration.hpp and Configuration_adv.hpp 
-#include "Configuration.hpp"
-#include "a_inits.hpp"
-#include "b_setup.hpp"
-#include "c65_startup.hpp"
-#include "c70_menuRA.hpp"
-#include "c71_menuDEC.hpp"
-#include "c722_menuPOI.hpp"
-#include "c725_menuHOME.hpp"
-#include "c72_menuHA.hpp"
-#include "c72_menuHA_GPS.hpp"
-#include "c74_menuHEAT.hpp"
-#include "c75_menuCTRL.hpp"
-#include "c76_menuCAL.hpp"
-#include "c78_menuINFO.hpp"
-#include "c_buttons.hpp"
-#include "f_serial.hpp"
-#include "g_bluetooth.hpp"
+
+// TODO: general OAT description
