+++ Software/Arduino code/OpenAstroTracker/platformio.ini
@@ -9,14 +9,13 @@
 ; https://docs.platformio.org/page/projectconf.html
 
 [platformio]
-src_dir = .
-default_envs = mega2560, esp32
+default_envs = mega2560
 
 [common]
 lib_deps = 
 	SPI
-	Serial
-	887 # LiquidCrystal
+	4625
+	887@1.0.7 # LiquidCrystal (set to 1.0.7 because arduino ide does not support newer)
 	265 # AccelStepper
 	EEPROM
 	Wire
