+++ Software/Arduino code/OpenAstroTracker/src/Core.cpp
@@ -1,0 +1,17 @@
+#include "../Configuration.hpp"
+#include "a_inits.hpp"
+#include "b_setup.hpp"
+#include "c65_startup.hpp"
+#include "c70_menuRA.hpp"
+#include "c71_menuDEC.hpp"
+#include "c722_menuPOI.hpp"
+#include "c725_menuHOME.hpp"
+#include "c72_menuHA.hpp"
+#include "c72_menuHA_GPS.hpp"
+#include "c74_menuHEAT.hpp"
+#include "c75_menuCTRL.hpp"
+#include "c76_menuCAL.hpp"
+#include "c78_menuINFO.hpp"
+#include "c_buttons.hpp"
+#include "f_serial.hpp"
+#include "g_bluetooth.hpp"
