+++ Software/Arduino code/OpenAstroTracker/src/DayTime.cpp
