+++ Software/Arduino code/OpenAstroTracker/src/InterruptCallback.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Configuration_adv.hpp"
+#include "../Configuration_adv.hpp"
 
 //////////////////////////////////////
 // This is an hardware-independent abstraction layer over
