+++ Software/Arduino code/OpenAstroTracker/src/MeadeCommandProcessor.cpp
@@ -1,5 +1,5 @@
 #include "MeadeCommandProcessor.hpp"
-#include "Configuration_adv.hpp"
+#include "../Configuration_adv.hpp"
 #include "Utility.hpp"
 #include "WifiControl.hpp"
