+++ Software/Arduino code/OpenAstroTracker/src/MeadeCommandProcessor.hpp
