+++ Software/Arduino code/OpenAstroTracker/src/Sidereal.cpp
