+++ Software/Arduino code/OpenAstroTracker/src/Sidereal.hpp
