+++ Software/Arduino code/OpenAstroTracker/src/Utility.cpp
