+++ Software/Arduino code/OpenAstroTracker/src/WifiControl.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "Configuration_adv.hpp"
+#include "../Configuration_adv.hpp"
 
 #ifdef WIFI_ENABLED
 #include "Mount.hpp"
