+++ Software/Arduino code/OpenAstroTracker/src/c76_menuCAL.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "Configuration_adv.hpp"
+#include "../Configuration_adv.hpp"
 #include "b_setup.hpp"
 
 #if HEADLESS_CLIENT == 0
