+++ Software/Arduino code/OpenAstroTracker/src/g_bluetooth.hpp
