+++ Software/Arduino code/OpenAstroTracker/src/libs/ESP8266_ISR/LICENSE
