+++ Software/Arduino code/OpenAstroTracker/src/libs/TimerInterrupt/LICENSE
